libxdp: Expose xdp_program__load()
We need to be able to load programs to run them through bpf_prog_run(), so
expose the xdp_program__load() helper.

This is probably not the right thing to do; we'll need to deal with the
program type replacement somehow, since now a program loaded through
xdp_program__load() won't be attachable.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>